add pre_process data
