--- paddlenlp/trainer/mutil_dataset_with_resume.py
@@ -1,2222 +1,0 @@
-import collections
-import contextlib
-import inspect
-import math
-import os
-import random
-import re
-import shutil
-import sys
-import time
-import types
-import warnings
-from collections import OrderedDict
-from collections.abc import Mapping
-from pathlib import Path
-from typing import Any, Callable, Dict, List, Optional, Tuple, Union
-
-import numpy as np
-import paddle
-import paddle.amp.auto_cast as autocast
-import paddle.distributed as dist
-import paddle.nn as nn
-from packaging import version
-from paddle.distributed import fleet
-from paddle.distributed.fleet.meta_optimizers.dygraph_optimizer.dygraph_sharding_optimizer import (
-    DygraphShardingOptimizer,
-)
-from paddle.distributed.fleet.meta_optimizers.dygraph_optimizer.hybrid_parallel_optimizer import (
-    HybridParallelOptimizer,
-)
-from paddle.distributed.fleet.meta_parallel.sharding.group_sharded_optimizer_stage2 import (
-    GroupShardedOptimizerStage2,
-)
-from .trainer import Trainer
-try:
-    from paddle.distributed.fleet.utils.hybrid_parallel_util import (
-        obtain_optimizer_parameters_list,
-    )
-
-    _obtain_optimizer_parameters_list = obtain_optimizer_parameters_list
-except:
-    try:
-        from paddle.distributed.fleet.meta_optimizers.dygraph_optimizer.hybrid_parallel_optimizer import (
-            _obtain_optimizer_parameters_list,
-        )
-    except:
-        _obtain_optimizer_parameters_list = None
-
-from paddle.distributed.fleet.utils.hybrid_parallel_util import (
-    fused_allreduce_gradients,
-)
-from paddle.io import DataLoader, Dataset, DistributedBatchSampler
-from tqdm.auto import tqdm
-
-from ..data import (
-    DataCollator,
-    DataCollatorWithPadding,
-    DistDataLoader,
-    default_data_collator,
-)
-from ..data.causal_dataset import get_datasets_weights_and_num_samples, build_train_valid_test_datasets
-from ..peft import LoRAModel, PrefixModelForCausalLM
-from ..transformers.model_utils import (
-    PretrainedModel,
-    _add_variant,
-    load_sharded_checkpoint,
-    unwrap_model,
-)
-from ..transformers.tokenizer_utils import PretrainedTokenizer
-from ..utils.batch_sampler import DistributedBatchSampler as NlpDistributedBatchSampler
-from ..utils.env import (
-    LORA_WEIGHTS_NAME,
-    PADDLE_WEIGHTS_INDEX_NAME,
-    PADDLE_WEIGHTS_NAME,
-    PREFIX_WEIGHTS_NAME,
-)
-from ..utils.import_utils import is_datasets_available
-from ..utils.log import logger
-from .integrations import get_reporting_integration_callbacks
-from .plugins.timer import get_timers, set_timers
-from .plugins.unified_checkpoint import load_unified_checkpoint, save_unified_checkpoint
-from .trainer_callback import (
-    CallbackHandler,
-    DefaultFlowCallback,
-    PrinterCallback,
-    ProgressCallback,
-    TrainerCallback,
-    TrainerControl,
-    TrainerState,
-)
-from .trainer_utils import (  # set_hyrbid_parallel_seed,
-    PREFIX_CHECKPOINT_DIR,
-    EvalLoopOutput,
-    EvalPrediction,
-    IterableDatasetShard,
-    OptimizerNames,
-    PredictionOutput,
-    RemoveColumnsCollator,
-    ShardingOption,
-    TrainerMemoryTracker,
-    TrainOutput,
-    find_batch_size,
-    get_last_checkpoint,
-    get_scheduler,
-    has_length,
-    set_seed,
-    speed_metrics,
-)
-from .training_args import TrainingArguments
-from .utils import reshard as reshard_util
-from .utils.helper import (  # nested_truncate,
-    broadcast_dp_optimizer,
-    distributed_concat,
-    distributed_file,
-    distributed_isfile,
-    nested_concat,
-    nested_detach,
-    nested_numpify,
-    nested_truncate,
-)
-from .utils.sharding_io import ShardingIO
-
-DEFAULT_CALLBACKS = [DefaultFlowCallback]
-DEFAULT_PROGRESS_CALLBACK = ProgressCallback
-
-# Name of the files used for checkpointing
-TRAINING_ARGS_NAME = "training_args.bin"
-TRAINER_STATE_NAME = "trainer_state.json"
-
-OPTIMIZER_NAME = "optimizer.pdopt"
-SCHEDULER_NAME = "scheduler.pdparams"
-SCALER_NAME = "scaler.pdparams"
-
-
-if is_datasets_available():
-    import datasets
-
-
-try:
-    from paddle.distributed.fleet.utils import mix_precision_utils
-except:
-    mix_precision_utils = None
-
-try:
-    from paddle.io.dataloader.dataloader_iter import _DataLoaderIterBase
-except:
-    from paddle.fluid.dataloader.dataloader_iter import _DataLoaderIterBase
-
-
-def is_dp_group_support_in_group_sharded_parallel():
-    return "dp_group" in set(inspect.signature(paddle.distributed.sharding.group_sharded_parallel).parameters.keys())
-
-
-__all__ = [
-    "Mutil_dataset_PretrainingTrainer",
-]
-
-class Mutil_dataset_PretrainingTrainer(Trainer):
-    def __init__(
-        self,
-        model: Union[PretrainedModel, nn.Layer] = None,
-        criterion: nn.Layer = None,
-        args: TrainingArguments = None,
-        data_collator: Optional[DataCollator] = None,
-        data: List = None,
-        data_args = None,
-        train_val_test_num_samples: List = None,
-        # train_dataset: Optional[Dataset] = None,
-        # eval_dataset: Union[Dataset, Dict[str, Dataset]] = None,
-        tokenizer: Optional[PretrainedTokenizer] = None,
-        compute_metrics: Optional[Callable[[EvalPrediction], Dict]] = None,
-        callbacks: Optional[List[TrainerCallback]] = None,
-        optimizers: Tuple[paddle.optimizer.Optimizer, paddle.optimizer.lr.LRScheduler] = (None, None),
-        preprocess_logits_for_metrics: Callable[[paddle.Tensor, paddle.Tensor], paddle.Tensor] = None,
-    ):
-        if args is None:
-            output_dir = "tmp_trainer"
-            logger.info(f"No `TrainingArguments` passed, using `output_dir={output_dir}`.")
-            args = TrainingArguments(output_dir=output_dir)
-
-        self.args = args
-        self.is_in_train = False
-        self.data = data
-        self.train_val_test_num_samples = train_val_test_num_samples
-        self.data_args = data_args
-        self.train_dataset = None
-        self.eval_dataset = None
-        # self.do_grad_scaling = args.fp16
-
-        # memory metrics - must set up as early as possible
-        self._memory_tracker = TrainerMemoryTracker(self.args.skip_memory_metrics)
-        self._memory_tracker.start()
-
-        # Seed must be set before instantiating the model when using model
-        set_seed(args=self.args)
-
-        if model is None:
-            raise RuntimeError("`Trainer` requires either a `model` or `model_init` argument")
-
-        if self.args.should_save or self.args.should_save_model_state:
-            os.makedirs(self.args.output_dir, exist_ok=True)
-
-        self.sharding = None
-        if len(args.sharding) > 0:
-            if args.local_rank == -1:
-                raise ValueError("Using sharding only works in distributed training.")
-            self.sharding = True
-
-        # init parallel env
-        if paddle.distributed.get_world_size() > 1:
-            if self.args.use_hybrid_parallel:
-                self.hcg = fleet.get_hybrid_communicate_group()
-                self.dp_group = self.hcg.get_data_parallel_group()
-                self.sharding_group = self.hcg.get_sharding_parallel_group()
-
-        default_collator = default_data_collator if tokenizer is None else DataCollatorWithPadding(tokenizer)
-
-        self.data_collator = data_collator if data_collator is not None else default_collator
-        # self.train_dataset = train_dataset
-        # self.eval_dataset = eval_dataset
-        self.tokenizer = tokenizer
-        if not args.skip_profile_timer:
-            set_timers()
-        self.timers = get_timers()
-
-        self.model_wrapped = model
-        self.model = model
-        self.criterion = criterion
-
-        self.compute_metrics = compute_metrics
-        self.preprocess_logits_for_metrics = preprocess_logits_for_metrics
-        self.optimizer, self.lr_scheduler = optimizers
-        # Label smoothing
-        # if self.args.label_smoothing_factor != 0:
-        #     self.label_smoother = LabelSmoother(epsilon=self.args.label_smoothing_factor)
-        # else:
-        self.label_smoother = None
-        self.state = TrainerState()
-        self.control = TrainerControl()
-        self._signature_columns = None
-        self.optimizer_grouped_parameters = None
-        self.sharding_io = None
-        if self.args.should_save_sharding_stage1_model or self.args.should_load_sharding_stage1_model:
-            self.sharding_io = ShardingIO(self.args, self.model, self.optimizer)
-
-        if self.sharding is not None and self.optimizer is not None:
-            raise RuntimeError(
-                "Passing `optimizers` is not allowed if sharding is enabled."
-                "You should subclass `Trainer` and override the `create_optimizer_and_scheduler` method."
-            )
-        if self.args.pipeline_parallel_degree > 1:
-            from paddle.distributed.fleet.meta_parallel import PipelineLayer
-
-            assert (isinstance(model, LoRAModel) and isinstance(model.model, PipelineLayer)) or isinstance(
-                model, PipelineLayer
-            ), "Only support pipeline parallel mode when model is PipelineLayer!!!"
-
-        default_callbacks = DEFAULT_CALLBACKS + get_reporting_integration_callbacks(self.args.report_to)
-        callbacks = default_callbacks if callbacks is None else default_callbacks + callbacks
-        self.callback_handler = CallbackHandler(
-            callbacks, self.model, self.tokenizer, self.optimizer, self.lr_scheduler
-        )
-        self.add_callback(PrinterCallback if self.args.disable_tqdm else DEFAULT_PROGRESS_CALLBACK)
-
-        # if train_dataset is not None and not isinstance(train_dataset, collections.abc.Sized) and args.max_steps <= 0:
-        #     raise ValueError("train_dataset does not implement __len__, max_steps has to be specified")
-
-        self.do_grad_scaling = False
-        self.enable_autocast_context_manager = False
-
-        if args.fp16 or args.bf16:
-            logger.info("Using half precision")
-            self.enable_autocast_context_manager = True
-            self.do_grad_scaling = True if args.fp16 else False
-            self.amp_dtype = "float16" if args.fp16 else "bfloat16"
-            # fix for load saved fp16 or bf16 ckpt, decorate model first.
-            if self.args.fp16_opt_level == "O2":
-                if self.amp_dtype == "bfloat16":
-                    # fix for paddlepaddle < 2.4.1, not support for bf16
-                    paddle.amp.decorate(models=model, level=self.args.fp16_opt_level, dtype=self.amp_dtype)
-                else:
-                    paddle.amp.decorate(models=model, level=self.args.fp16_opt_level)
-            # for pipeline mode and pure tensor parallel
-            if self.args.pipeline_parallel_degree > 1 or (
-                self.args.tensor_parallel_degree > 1 and self.sharding is None
-            ):
-                self.scaler = paddle.amp.GradScaler(init_loss_scaling=self.args.scale_loss)
-                if self.args.amp_master_grad:
-                    mix_precision_utils.MixPrecisionScaler(self.scaler)  # retun value has no use
-                self.scaler = fleet.distributed_scaler(self.scaler)
-            elif self.sharding is not None:
-                self.scaler = paddle.amp.GradScaler(init_loss_scaling=self.args.scale_loss)
-                if self.amp_dtype == "float16" or self.amp_dtype == "bfloat16":
-                    if ShardingOption.SHARD_OP in self.args.sharding:
-                        self.scaler = fleet.distributed_scaler(self.scaler)
-                        if self.args.amp_master_grad:
-                            mix_precision_utils.MixPrecisionScaler(self.scaler)  # retun value has no use
-                    else:
-                        # scaler for stage2 and stage3
-                        from paddle.distributed.fleet.meta_parallel.sharding.group_sharded_utils import (
-                            GroupShardedScaler,
-                        )
-
-                        if self.args.amp_master_grad:
-                            assert (
-                                ShardingOption.SHARD_GRAD_OP in self.args.sharding
-                            ), "Main grad doesn't support sharding stage 3 for now."
-                            mix_precision_utils.MixPrecisionScaler(self.scaler)  # return value has no use
-
-                        self.scaler = GroupShardedScaler(self.scaler)
-                else:
-                    self.do_grad_scaling = False
-                    self.use_cuda_amp = False
-                    self.amp_dtype = None
-
-            else:
-                self.scaler = paddle.amp.GradScaler(init_loss_scaling=self.args.scale_loss)
-
-        if args.recompute:
-
-            def fn(layer):
-                if hasattr(layer, "enable_recompute") and (
-                    layer.enable_recompute is False or layer.enable_recompute == 0
-                ):
-                    layer.enable_recompute = True
-
-            model.apply(fn)
-
-        default_label_names = (
-            ["start_positions", "end_positions"]
-            if "QusetionAnswering" in type(self.model).__name__ or "UIE" in type(self.model).__name__
-            else ["labels"]
-        )
-        self.label_names = default_label_names if self.args.label_names is None else self.args.label_names
-
-        self.control = self.callback_handler.on_init_end(self.args, self.state, self.control)
-        self.print_config()
-
-        self._memory_tracker.stop_and_update_metrics()
-
-    # 怎么选择dataset
-    def get_dataset(self, datasets_train_valid_test_num_samples, total_batch_size, datasets_total_samples_dict):
-        datasets_need_samples = [dataset[0] for dataset in datasets_train_valid_test_num_samples]
-        datasets_steps = [max(1, samples // total_batch_size) for samples in datasets_need_samples]
-        sum_steps = sum(datasets_steps)
-        map_dataset = np.arange(sum_steps, dtype=np.int8)
-        if datasets_total_samples_dict is None:
-            return map_dataset
-        else:
-            datasets_total_samples = list(datasets_total_samples_dict.values())
-            datasets_total_step = [max(1, dataset_total_samples // total_batch_size) for dataset_total_samples in datasets_total_samples]
-            datasets_need_step = []
-            for i, dataset_total_step in enumerate(datasets_total_step):
-                datasets_need_step.append(min(dataset_total_step, datasets_steps[i]))
-            offset = 0
-            for index, dataset_need_step in enumerate(datasets_need_step):
-                map_dataset[offset:offset+dataset_need_step] = [index] * len(map_dataset[offset:offset+dataset_need_step])
-                offset += dataset_need_step
-            if len(datasets_train_valid_test_num_samples) == len(datasets_need_step):
-                left_steps = [i-j for i, j in zip(datasets_steps, datasets_need_step)]
-                while sum(left_steps) > 0:
-                    left_steps_2 = []
-                    for i, left_step in enumerate(left_steps):
-                        left_steps_2.append(min(left_step, datasets_total_step[i]))
-                    for index, left_step_2 in enumerate(left_steps_2):
-                        map_dataset[offset:offset+left_step_2] = [index] * len(map_dataset[offset:offset+left_step_2])
-                        offset += left_step_2
-                    left_steps = [i-j for i, j in zip(left_steps, left_steps_2)]
-            else:
-                map_dataset[offset] = index + 1
-        return map_dataset
-    
-    def get_offset_step(self, step, datasets_train_valid_test_num_samples, total_batch_size, datasets_total_samples_dict):
-        datasets_total_samples = list(datasets_total_samples_dict.values())
-        datasets_need_samples = [dataset[0] for dataset in datasets_train_valid_test_num_samples]
-        datasets_steps = [max(1, samples // total_batch_size) for samples in datasets_need_samples]
-        sum_steps = sum(datasets_steps)
-        offset_step = np.arange(sum_steps, dtype=np.int8)
-        if len(datasets_total_samples) == 1:
-            return offset_step[step]
-        else:
-            datasets_total_step = [max(1, dataset_total_samples // total_batch_size) for dataset_total_samples in datasets_total_samples]
-            datasets_need_step = []
-            for i, dataset_total_step in enumerate(datasets_total_step):
-                datasets_need_step.append(min(dataset_total_step, datasets_steps[i]))
-            offset = 0
-            for index, dataset_need_step in enumerate(datasets_need_step):
-                offset_step[offset:offset+dataset_need_step] = np.arange(dataset_need_step)
-                offset += dataset_need_step
-            if len(datasets_train_valid_test_num_samples) == len(datasets_need_step):
-                left_steps = [i-j for i, j in zip(datasets_steps, datasets_need_step)]
-                while sum(left_steps) > 0:
-                    left_steps_2 = []
-                    for i, left_step in enumerate(left_steps):
-                        left_steps_2.append(min(left_step, datasets_total_step[i]))
-                    for left_step_2 in left_steps_2:
-                        offset_step[offset:offset+left_step_2] = np.arange(left_step_2)
-                        offset += left_step_2
-                    left_steps = [i-j for i, j in zip(left_steps, left_steps_2)]
-            else:
-                offset_step[offset] = 0
-        return offset_step[step]
-
-    def add_callback(self, callback):
-        """
-        Add a callback to the current list of [`~TrainerCallback`].
-
-        Args:
-           callback (`type` or [`~TrainerCallback`]):
-               A [`~TrainerCallback`] class or an instance of a [`~TrainerCallback`]. In the
-               first case, will instantiate a member of that class.
-        """
-        self.callback_handler.add_callback(callback)
-
-    def pop_callback(self, callback):
-        """
-        Remove a callback from the current list of [`~TrainerCallback`] and returns it.
-        If the callback is not found, returns `None` (and no error is raised).
-        Args:
-           callback (`type` or [`~TrainerCallback`]):
-               A [`~TrainerCallback`] class or an instance of a [`~TrainerCallback`]. In the
-               first case, will pop the first member of that class found in the list of callbacks.
-        Returns:
-            [`~TrainerCallback`]: The callback removed, if found.
-        """
-        return self.callback_handler.pop_callback(callback)
-
-    def remove_callback(self, callback):
-        """
-        Remove a callback from the current list of [`~TrainerCallback`].
-        Args:
-           callback (`type` or [`~TrainerCallback`]):
-               A [`~TrainerCallback`] class or an instance of a [`~TrainerCallback`]. In the
-               first case, will remove the first member of that class found in the list of callbacks.
-        """
-        self.callback_handler.remove_callback(callback)
-
-    def _load_from_peft_checkpoint(self, resume_from_checkpoint=None):
-        """load state_dict from checkpoint, Only for PEFT Model.
-
-        Args:
-            resume_from_checkpoint (`str` or `bool`, *optional*):
-                If a `str`, local path to a saved checkpoint as saved by a previous instance of [`Trainer`]. If a
-                `bool` and equals `True`, load the last checkpoint in *args.output_dir* as saved by a previous instance
-                of [`Trainer`]. Only load model state dict.
-        """
-
-        if resume_from_checkpoint is not None:
-            convert_tp = False
-            if isinstance(self.model, LoRAModel):
-                if self.model.quantized or self.args.pipeline_parallel_degree > 1:
-                    weights_file = os.path.join(
-                        resume_from_checkpoint, _add_variant(LORA_WEIGHTS_NAME, self.args.weight_name_suffix)
-                    )
-                else:
-                    weights_file = os.path.join(resume_from_checkpoint, LORA_WEIGHTS_NAME)
-                    if self.model.lora_config.tensor_parallel_degree > 1:
-                        convert_tp = True
-            elif isinstance(self.model, PrefixModelForCausalLM):
-                weights_file = os.path.join(resume_from_checkpoint, PREFIX_WEIGHTS_NAME)
-                if self.model.prefix_config.tensor_parallel_degree > 1:
-                    convert_tp = True
-            if self.args.dataset_rank == 0:
-                logger.info(f"Loading model from {resume_from_checkpoint} .")
-
-                if os.path.isfile(weights_file):
-                    # We load the model state dict on the CPU to avoid an OOM error.
-                    state_dict = paddle.load(weights_file, return_numpy=True)
-                    if convert_tp:
-                        state_dict = self.model._convert_tensor_parallel(state_dict)
-
-                    # If the model is on the GPU, it still works!
-                    self._set_state_dict_in_model(state_dict)
-                    # release memory
-                    del state_dict
-        elif resume_from_checkpoint is not None:
-            logger.info(f"not loading ckpt :{self.args.dataset_rank}")
-
-    def _load_from_checkpoint(self, resume_from_checkpoint=None):
-        """load state_dict from_checkpoint, Only load model state dict.
-
-        Args:
-            resume_from_checkpoint (`str` or `bool`, *optional*):
-                If a `str`, local path to a saved checkpoint as saved by a previous instance of [`Trainer`]. If a
-                `bool` and equals `True`, load the last checkpoint in *args.output_dir* as saved by a previous instance
-                of [`Trainer`]. Only load model state dict.
-        """
-        resume_from_checkpoint = None if not resume_from_checkpoint else resume_from_checkpoint
-
-        # Load potential model checkpoint
-        if isinstance(resume_from_checkpoint, bool) and resume_from_checkpoint:
-            resume_from_checkpoint = get_last_checkpoint(self.args.output_dir)
-            if resume_from_checkpoint is None:
-                raise ValueError(f"No valid checkpoint found in output directory ({self.args.output_dir})")
-
-        if self.args.unified_checkpoint:
-            if resume_from_checkpoint is not None and self.args.dataset_rank == 0:
-                load_unified_checkpoint(
-                    self.model,
-                    resume_from_checkpoint,
-                    safe_serialization=True,
-                )
-                logger.info(f"Loading model from {resume_from_checkpoint} using unified checkpoint.")
-                return
-
-        if isinstance(self.model, LoRAModel) or isinstance(self.model, PrefixModelForCausalLM):
-            self._load_from_peft_checkpoint(resume_from_checkpoint)
-            return
-
-        weight_name = PADDLE_WEIGHTS_NAME
-        weight_index_name = PADDLE_WEIGHTS_INDEX_NAME  # currently set paddle as default, do not support safetensors.
-
-        if self.args.should_load_sharding_stage1_model:
-            state_dict = self.sharding_io.load_state_dict_from_checkpoint_with_reshard(
-                resume_from_checkpoint,
-                base_weight_name=weight_name,
-            )
-        else:
-            if resume_from_checkpoint is not None and self.args.dataset_rank == 0:
-
-                weights_file = os.path.join(
-                    resume_from_checkpoint, _add_variant(weight_name, self.args.weight_name_suffix)
-                )
-                weights_index_file = os.path.join(
-                    resume_from_checkpoint, _add_variant(weight_index_name, self.args.weight_name_suffix)
-                )
-
-                if not any(
-                    os.path.isfile(f)
-                    for f in [
-                        weights_file,
-                        weights_index_file,
-                    ]
-                ):
-                    raise ValueError(f"Can't find a valid checkpoint at {resume_from_checkpoint}")
-
-                logger.info(f"Loading model from {resume_from_checkpoint} .")
-
-                if os.path.isfile(weights_file):
-                    # We load the model state dict on the CPU to avoid an OOM error.
-                    state_dict = paddle.load(weights_file, return_numpy=True)
-
-                    # If the model is on the GPU, it still works!
-                    self._set_state_dict_in_model(state_dict)
-                    # release memory
-                    del state_dict
-                else:
-                    # We load the sharded checkpoint.
-                    missing_keys, unexpected_keys = load_sharded_checkpoint(
-                        self.model, resume_from_checkpoint, self.args.weight_name_suffix, prefer_safe=False
-                    )
-                    logger.info(f"set state_dict: {missing_keys, unexpected_keys}")
-
-            elif resume_from_checkpoint is not None:
-                logger.info(f"not loading ckpt :{self.args.dataset_rank}")
-
-    def _wrap_model_and_load_sharded_checkpoint(self, resume_from_checkpoint):
-        # In the sharded mode, should invoke _load_from_checkpoint after _wrap_model.
-        # In this mode, each sharding rank load sharded params, do not need to implement the broadcast logic.
-        model = self._wrap_model(self.model_wrapped)
-        if self.sharding_io is not None:
-            # the self.optimizer should be wrapped and it is done in _wrap_model
-            self.sharding_io.set_optimizer(self.optimizer)
-        if model is not self.model:
-            self.model_wrapped = model
-        # Should invoke _load_from_checpoint after _load_optimizer_and_scheduler
-        # because the _load_from_checkpoint method rely on the optimizer in the shareded mode.
-        if resume_from_checkpoint:
-            self._load_optimizer_and_scheduler(resume_from_checkpoint)
-            self._load_from_checkpoint(resume_from_checkpoint)
-        return model
-
-    def train(
-        self,
-        resume_from_checkpoint: Optional[Union[str, bool]] = None,
-        ignore_keys_for_eval: Optional[List[str]] = None,
-    ):
-        """
-        Main training entry point.
-
-        Args:
-            resume_from_checkpoint (`str` or `bool`, *optional*):
-                If a `str`, local path to a saved checkpoint as saved by a previous instance of [`Trainer`]. If a
-                `bool` and equals `True`, load the last checkpoint in *args.output_dir* as saved by a previous instance
-                of [`Trainer`]. If present, training will resume from the model/optimizer/scheduler states loaded here.
-            ignore_keys_for_eval (`List[str]`, *optional*)
-                A list of keys in the output of your model (if it is a dictionary) that should be ignored when
-                gathering predictions for evaluation during the training.
-        """
-        args = self.args
-        self.is_in_train = True
-
-        logger.info(f"Starting training from resume_from_checkpoint : {resume_from_checkpoint}")
-
-        # The resume_from_checkpoint could be None in some machine node.
-        # Here we reset None to temp directory.
-        if args.world_size > 1:
-            is_resume_from_checkpoint = paddle.to_tensor([resume_from_checkpoint is not None])
-            paddle.distributed.all_reduce(is_resume_from_checkpoint)
-            is_resume_from_checkpoint = is_resume_from_checkpoint.item()
-            if is_resume_from_checkpoint > 0 and is_resume_from_checkpoint < paddle.distributed.get_world_size():
-                if resume_from_checkpoint is None:
-                    resume_from_checkpoint = os.path.join(self.args.output_dir, "local_tempdir")
-                    if os.path.exists(resume_from_checkpoint) and self.args.local_rank == 0:
-                        shutil.rmtree(resume_from_checkpoint)
-                    os.makedirs(resume_from_checkpoint, exist_ok=True)
-                    logger.info(f"Reset resume_from_checkpoint to temp directory : {resume_from_checkpoint}")
-
-        # memory metrics - must set up as early as possible
-        self._memory_tracker.start()
-
-        if not self.args.should_load_sharding_stage1_model:
-            self._load_from_checkpoint(resume_from_checkpoint)
-
-        total_train_batch_size = args.train_batch_size * args.gradient_accumulation_steps * args.dataset_world_size
-        output = get_datasets_weights_and_num_samples(self.data, self.train_val_test_num_samples)
-        prefixes, weights, datasets_train_valid_test_num_samples = output
-        train_num_samples, valid_num_samples, test_num_samples = map(sum, zip(*datasets_train_valid_test_num_samples))
-        
-        max_steps = args.max_steps
-        # num_train_samples = train_num_samples
-        # delay_optimizer_creation = (
-        #     self.sharding is not None
-        #     and ShardingOption.SHARD_OP in self.args.sharding
-        # )
-        delay_optimizer_creation = False
-
-        if not delay_optimizer_creation:
-            self.create_optimizer_and_scheduler(num_training_steps=max_steps)
-
-        self.state = TrainerState()
-
-        if self.args.should_load_sharding_stage1_model:
-            model = self._wrap_model_and_load_sharded_checkpoint(resume_from_checkpoint)
-        elif self.args.should_save_sharding_stage1_model:
-            # In the non-sharded mode, should invoke _load_from_checkpoint before _wrap_model.
-            # In this mode, the rank0 load all params and the _wrap_model implicitly broadcast params from rank0 to the other ranks.
-            model = self._wrap_model(self.model_wrapped)
-            if self.sharding_io is not None:
-                assert delay_optimizer_creation is False, "delay_optimizer_creation should be False"
-                # the self.optimizer should be wrapped and it is done in _wrap_model
-                self.sharding_io.set_optimizer(self.optimizer)
-            # for the rest of this function `model` is the outside model, whether it was wrapped or not
-            if model is not self.model:
-                self.model_wrapped = model
-            if delay_optimizer_creation:
-                self.create_optimizer_and_scheduler(num_training_steps=max_steps)
-            self._load_optimizer_and_scheduler(resume_from_checkpoint)
-        else:
-            model = self._wrap_model(self.model_wrapped)
-            # for the rest of this function `model` is the outside model, whether it was wrapped or not
-            if model is not self.model:
-                self.model_wrapped = model
-            if delay_optimizer_creation:
-                self.create_optimizer_and_scheduler(num_training_steps=max_steps)
-            self._load_optimizer_and_scheduler(resume_from_checkpoint)
-
-        logger.info("***** Running training *****")
-        logger.info(f"  Instantaneous batch size per device = {args.per_device_train_batch_size}")
-        logger.info(f"  Total train batch size (w. parallel, distributed & accumulation) = {total_train_batch_size}")
-        logger.info(f"  Gradient Accumulation steps = {args.gradient_accumulation_steps}")
-        logger.info(f"  Total optimization steps = {max_steps:,}")
-        logger.info(f"  Total num train samples = {train_num_samples:,}")
-
-        # per_device_trainable_numel = sum(p.numel().item() for p in model.parameters() if not p.stop_gradient)
-        # TODO: Temporary fix since Tensor.numel() not supported in distributed mode
-        per_device_trainable_numel = sum(np.prod(p.shape) for p in model.parameters() if not p.stop_gradient)
-        logger.info(f"  Number of trainable parameters = {per_device_trainable_numel:,} (per device)")
-        if self.args.use_hybrid_parallel:
-            # todo fix for pipeline_parallel_degree
-            parts_num = max(self.args.tensor_parallel_degree, 1) * max(self.args.pipeline_parallel_degree, 1)
-            if parts_num > 1:
-                all_reduce_dtype = "int64"
-                if paddle.get_device().split(":")[0] in ["npu", "xpu"]:
-                    # TODO(duanyanhui): fix when NPU all_reduce supports int64
-                    all_reduce_dtype = "float32"
-                trainable_numel_tensor = paddle.to_tensor(per_device_trainable_numel, dtype=all_reduce_dtype)
-                paddle.distributed.all_reduce(trainable_numel_tensor)
-                trainable_numel = int(trainable_numel_tensor.item()) // self.args.dataset_world_size
-                # the numel is roughly, because the tensor parallel still hold own bias or layer_norm weight without splited
-                # so, the trainable numel is a little bigger than real.
-                logger.info(f"  Number of trainable parameters = {trainable_numel:,} (all devices, roughly)")
-
-        start_time = time.time()
-        self._globalstep_last_start_time = time.time()
-
-        # Check if continuing training from a checkpoint
-        if resume_from_checkpoint is not None and distributed_isfile(
-            os.path.join(resume_from_checkpoint, TRAINER_STATE_NAME)
-        ):
-            self.state = TrainerState.load_from_json(
-                distributed_file(os.path.join(resume_from_checkpoint, TRAINER_STATE_NAME))
-            )
-            self._load_rng_state(resume_from_checkpoint)
-            if self.args.world_size > 1:
-                global_step_list = []
-                paddle.distributed.all_gather(
-                    global_step_list, paddle.to_tensor([self.state.global_step], dtype="int64")
-                )
-                assert (
-                    paddle.sum(paddle.stack(global_step_list) - global_step_list[0]) == 0
-                ), f"Error, get different globel step, please check! step list: {[x.item() for x in global_step_list]}"
-
-            logger.info("  Continuing training from checkpoint, will skip to saved global_step")
-            logger.info(f"  Continuing training from global step {self.state.global_step}")
-
-        self.callback_handler.model = self.model
-        self.callback_handler.optimizer = self.optimizer
-        self.callback_handler.lr_scheduler = self.lr_scheduler
-
-        self.state.max_steps = int(max_steps)
-        self.state.is_local_process_zero = self.is_local_process_zero()
-        self.state.is_world_process_zero = self.is_world_process_zero()
-
-        self.control = self.callback_handler.on_train_begin(args, self.state, self.control)
-
-        tr_loss = paddle.to_tensor(0.0)
-        self._total_loss_scalar = 0.0
-        self._globalstep_last_logged = self.state.global_step
-
-        if self.args.device == "npu" and self.args.flatten_param_grads:
-            from .plugins.npu_plugin import npu_accelerate_plugin
-
-            npu_accelerate_plugin(self.optimizer)
-
-        front_dataset_id = -1
-        flag = 1
-        while self.state.global_step < max_steps:
-            self.timers and self.timers("read-dataset").start()
-            dataset_id_map = self.get_dataset(datasets_train_valid_test_num_samples, total_train_batch_size, self.state.dataset_samples)
-            dataset_id = dataset_id_map[self.state.global_step]
-            if front_dataset_id != dataset_id or self.train_dataset is None:
-                if self.train_dataset is not None:
-                    del train_dataset
-                if self.eval_dataset is not None:
-                    del val_dataset
-                train_dataset, val_dataset, _ = build_train_valid_test_datasets(
-                    [self.data[dataset_id*2+1]], 
-                    self.data_args.data_impl,
-                    self.data_args.split,
-                    self.train_val_test_num_samples,
-                    self.data_args.max_seq_length,
-                    self.args.seed,
-                    self.data_args.skip_warmup,
-                    share_folder=self.data_args.share_folder,
-                    data_cache_path=self.data_args.data_cache,
-                )
-                self.train_dataset = train_dataset
-                self.eval_dataset = val_dataset
-                skip_step = 0
-                if self.state.dataset_samples is None:
-                    self.state.dataset_samples = dict()
-                    save_name = self.data[dataset_id*2+1]
-                    self.state.dataset_samples[save_name] = len(train_dataset)
-                else:
-                    save_name = self.data[dataset_id*2+1]
-                    self.state.dataset_samples[save_name] = len(train_dataset)
-                train_dataloader = self.get_train_dataloader()
-                if isinstance(train_dataloader, paddle.io.DataLoader) and isinstance(
-                    train_dataloader.batch_sampler, DistributedBatchSampler
-                ):
-                    train_dataloader.batch_sampler.set_epoch(self.state.num_train_epochs)
-                self.callback_handler.train_dataloader = train_dataloader
-                if flag == 1 and resume_from_checkpoint is not None and distributed_isfile(os.path.join(resume_from_checkpoint, TRAINER_STATE_NAME)):
-                    skip_step = self.get_offset_step(self.state.global_step, datasets_train_valid_test_num_samples, total_train_batch_size, self.state.dataset_samples)
-                    if isinstance(train_dataloader, paddle.io.DataLoader) and isinstance(
-                        train_dataloader.batch_sampler, NlpDistributedBatchSampler
-                    ):
-                        consumed_samples = (
-                            skip_step
-                            * args.train_batch_size
-                            * args.gradient_accumulation_steps
-                            * args.dataset_world_size
-                        )
-                        train_dataloader.batch_sampler.set_epoch(consumed_samples=consumed_samples)
-                        logger.info(f"Set DistributedBatchSampler consumed_samples to {consumed_samples}")
-                        flag = 0
-                
-            else:
-                if isinstance(train_dataloader, paddle.io.DataLoader) and isinstance(
-                    train_dataloader.batch_sampler, DistributedBatchSampler
-                ):
-                    train_dataloader.batch_sampler.set_epoch(self.state.num_train_epochs)
-            self.timers and self.timers("read-dataset").stop()
-            front_dataset_id = dataset_id
-            self.control = self.callback_handler.on_epoch_begin(args, self.state, self.control)
-            self.timers and self.timers("read-data").start()
-            consumed_samples = (
-                skip_step
-                * args.train_batch_size
-                * args.gradient_accumulation_steps
-                * args.dataset_world_size
-            )
-            for step, inputs in enumerate(train_dataloader):
-                if flag == 1 and resume_from_checkpoint is not None and distributed_isfile(os.path.join(resume_from_checkpoint, TRAINER_STATE_NAME)) and isinstance(train_dataloader.batch_sampler, DistributedBatchSampler):
-                    if skip_step > 0:
-                        skip_step -= 1
-                        continue
-                    else:
-                        logger.info(f"Set DistributedBatchSampler consumed_samples to {consumed_samples}")
-                        flag = 0
-                print(inputs)
-                self.timers and self.timers("read-data").stop()
-                os.environ["TRAINER_GLOBAL_STEP"] = str(self.state.global_step)
-                self.callback_handler.on_load_data_end(args, self.state, self.control, inputs=inputs)
-
-                if self.state.global_step % args.gradient_accumulation_steps == 0:
-                    self.control = self.callback_handler.on_step_begin(args, self.state, self.control)
-                    self.timers and self.timers("forward-backward").start()
-
-                dp_enabled = (
-                    self.args.data_parallel_degree > 1 if self.args.use_hybrid_parallel else args.local_rank != -1
-                )
-                forbidden_no_sync = False
-                # stage2 and stage3 should not no_sync, because the is no DDP wrapper and no_sync API
-                # hybrid_parallel (tp or pp or sharding stage 1) should not no_sync
-                if self.args.use_hybrid_parallel:
-                    forbidden_no_sync = True
-
-                availiable_no_sync = dp_enabled and not forbidden_no_sync
-
-                is_no_sync = (
-                    ((self.state.global_step + 1) % args.gradient_accumulation_steps != 0)
-                    and availiable_no_sync
-                    and args._no_sync_in_gradient_accumulation
-                ) or (args.recompute and availiable_no_sync)
-                # sharding
-                # stage1. the same as ddp
-                # stage2. manualy collect gradient on dp group
-                if is_no_sync:
-                    # Avoid unnecessary DDP synchronization since there will be no backward pass on this example.
-                    with model.no_sync():
-                        tr_loss_step = self.training_step(model, inputs)
-                else:
-                    tr_loss_step = self.training_step(model, inputs)
-
-                tr_loss += tr_loss_step
-
-                if (self.state.global_step + 1) % args.gradient_accumulation_steps == 0 or (
-                    # last step in epoch but step is always smaller than gradient_accumulation_steps
-                    (self.state.global_step + 1) % args.gradient_accumulation_steps != 0
-                    and (self.state.global_step + 1) == max_steps
-                ):
-                    self.timers and self.timers("forward-backward").stop()
-                    # Maunally collect gradients when group_sharded_parallel can't accept dp_group
-                    # Case 1: Use sharding stage 2/3 with dp
-                    # Case 2: Use recompute and dp
-                    # local_rank != -1 don't means dp in networks.
-                    self.timers and self.timers("all-reduce").start()
-
-                    if self.sharding and ShardingOption.SHARD_OP not in self.args.sharding:
-                        if self.args.data_parallel_degree > 1 and not is_dp_group_support_in_group_sharded_parallel():
-                            fused_allreduce_gradients(model.parameters(), fleet.get_hybrid_communicate_group())
-                            if ShardingOption.FULL_SHARD in self.args.sharding:
-                                # Why need sync on parm again ?
-                                # TODO: fix this.
-                                for p in model.parameters():
-                                    if hasattr(p, "bw_storage"):
-                                        assert p.grad is None, "This case shouldn't happen."
-                                        p.bw_storage.scale_(1.0 / self.dp_group.nranks)
-                                        paddle.distributed.all_reduce(p.bw_storage, group=self.dp_group)
-
-                    # Case 2: Use recompute and dp / sharding stage1,
-                    # manualy collect gradient for dp.
-                    elif args.recompute and availiable_no_sync:
-                        fused_allreduce_gradients(list(model.parameters()), None)
-
-                    pipeline_parallel_config = set(args.pipeline_parallel_config.split(" "))
-                    enable_delay_scale_loss = "enable_delay_scale_loss" in pipeline_parallel_config
-                    enable_dp_comm_overlap = "enable_dp_comm_overlap" in pipeline_parallel_config
-
-                    if isinstance(self.optimizer, HybridParallelOptimizer) and not self.do_grad_scaling:
-                        parameters_list = _obtain_optimizer_parameters_list(self.optimizer._inner_opt)
-
-                        if not enable_dp_comm_overlap:
-                            if self.optimizer._sharding_enable:
-                                assert reshard_util.is_sharding_opt(self.optimizer)
-                                self.optimizer._inner_opt.reduce_gradients(list(parameters_list), self.optimizer._hcg)
-
-                            if self.optimizer._dp_enable:
-                                fused_allreduce_gradients(list(parameters_list), self.optimizer._hcg)
-                    self.timers and self.timers("all-reduce").stop()
-                    self.timers and self.timers("optimizer-step").start()
-
-                    # pipeline parallel mode,  handle gradient merge here
-                    if args.pipeline_parallel_degree > 1 and enable_delay_scale_loss:
-                        for p in model._layers.parameters():
-                            with paddle.no_grad():
-                                if hasattr(p, "main_grad") and p.main_grad is not None:
-                                    assert p.grad is None
-                                    p.main_grad.scale_(1.0 / self.args.gradient_accumulation_steps)
-                                elif p.grad is not None:
-                                    p.grad.scale_(1.0 / self.args.gradient_accumulation_steps)
-
-                    # Optimizer step
-                    self.callback_handler.on_optimizer_begin(
-                        args, self.state, self.control, scaler=self.scaler if self.do_grad_scaling else None
-                    )
-                    optimizer_was_run = True
-                    if self.do_grad_scaling:
-                        scale_before = paddle.assign(self.scaler._scale)
-                        self.scaler.step(self.optimizer)
-                        self.scaler.update()
-                        scale_after = self.scaler._scale
-                        optimizer_was_run = not self.scaler._cache_founf_inf
-                        if not optimizer_was_run:
-                            scale_before_value = scale_before.cpu().numpy()
-                            scale_after_value = scale_after.cpu().numpy()
-                            logger.warning(
-                                f"optimizer not run, scale_before: {scale_before_value[0]}, scale_after: {scale_after_value[0]}"
-                            )
-                    elif isinstance(self.optimizer, HybridParallelOptimizer):
-                        self.optimizer._step(parameters_list)
-                    else:
-                        self.optimizer.step()
-
-                    self.timers and self.timers("optimizer-step").stop()
-
-                    if optimizer_was_run:
-                        self.lr_scheduler.step()
-
-                    self.optimizer.clear_grad()
-                    self.callback_handler.on_optimizer_end(
-                        args, self.state, self.control, scaler=self.scaler if self.do_grad_scaling else None
-                    )
-
-                    self.state.global_step += 1
-                    self.state.epoch = 0
-                    self.control = self.callback_handler.on_step_end(args, self.state, self.control)
-                    self._maybe_log_save_evaluate(tr_loss, model, 0, ignore_keys_for_eval, inputs=inputs)
-                    self._print_timer()
-                else:
-                    self.control = self.callback_handler.on_substep_end(args, self.state, self.control)
-
-                if self.state.global_step >= max_steps:
-                    break
-                if self.get_dataset(datasets_train_valid_test_num_samples, total_train_batch_size, self.state.dataset_samples)[self.state.global_step] != front_dataset_id:
-                    break
-                self.timers and self.timers("read-data").start()
-
-            self.state.num_train_epochs += 1
-
-            if step < 0:
-                logger.warning(
-                    f"There seems to be not a single sample in your epoch_iterator, stopping training at step"
-                    f" {self.state.global_step}! This is expected if you're using an IterableDataset and set"
-                    f" num_steps ({self.state.max_steps}) higher than the number of available samples."
-                )
-                self.control.should_training_stop = True
-
-            self.control = self.callback_handler.on_epoch_end(args, self.state, self.control)
-            self._maybe_log_save_evaluate(tr_loss, model, 0, ignore_keys_for_eval, inputs=inputs)
-
-            if self.control.should_training_stop:
-                break
-
-        if args.past_index and hasattr(self, "_past"):
-            # Clean the state at the end of training
-            delattr(self, "_past")
-
-        logger.info("\nTraining completed. \n")
-        if args.load_best_model_at_end and self.state.best_model_checkpoint is not None:
-            if args.local_rank != -1:
-                dist.barrier()
-
-            logger.info(
-                f"Loading best model from {self.state.best_model_checkpoint} (score: {self.state.best_metric})."
-            )
-            if isinstance(self.model, LoRAModel) or isinstance(self.model, PrefixModelForCausalLM):
-                self._load_best_model_from_peft_checkpoint()
-            else:
-                weight_name = PADDLE_WEIGHTS_NAME
-                best_model_path = os.path.join(
-                    self.state.best_model_checkpoint, _add_variant(weight_name, self.args.weight_name_suffix)
-                )
-                if os.path.exists(best_model_path):
-                    # We load the model state dict on the CPU to avoid an OOM error.
-                    state_dict = paddle.load(best_model_path, return_numpy=True)
-                    # If the model is on the GPU, it still works!
-                    self._set_state_dict_in_model(state_dict)
-                else:
-                    logger.warning(
-                        f"Could not locate the best model at {best_model_path}, if you are running a distributed training "
-                        "on multiple nodes, you should activate `--save_on_each_node`."
-                    )
-
-        self._total_loss_scalar += tr_loss.item()
-        train_loss = self._total_loss_scalar / self.state.global_step
-
-        metrics = speed_metrics("train", start_time, num_samples=train_num_samples, num_steps=self.state.max_steps)
-
-        metrics["train_loss"] = train_loss
-
-        self.is_in_train = False
-
-        self._memory_tracker.stop_and_update_metrics(metrics)
-
-        self.log(metrics)
-
-        self.control = self.callback_handler.on_train_end(args, self.state, self.control)
-
-        return TrainOutput(self.state.global_step, train_loss, metrics)
-
-    def _load_best_model_from_peft_checkpoint(self):
-        convert_tp = False
-        if isinstance(self.model, LoRAModel):
-            if self.model.quantized or self.args.pipeline_parallel_degree > 1:
-                best_model_path = os.path.join(
-                    self.state.best_model_checkpoint, _add_variant(LORA_WEIGHTS_NAME, self.args.weight_name_suffix)
-                )
-            else:
-                best_model_path = os.path.join(self.state.best_model_checkpoint, LORA_WEIGHTS_NAME)
-                if self.model.lora_config.tensor_parallel_degree > 1:
-                    convert_tp = True
-
-        elif isinstance(self.model, PrefixModelForCausalLM):
-            best_model_path = os.path.join(self.state.best_model_checkpoint, PREFIX_WEIGHTS_NAME)
-            if self.model.prefix_config.tensor_parallel_degree > 1:
-                convert_tp = True
-
-        if os.path.exists(best_model_path):
-            # We load the model state dict on the CPU to avoid an OOM error.
-            state_dict = paddle.load(best_model_path, return_numpy=True)
-            if convert_tp:
-                state_dict = self.model._convert_tensor_parallel(state_dict)
-            # If the model is on the GPU, it still works!
-            self._set_state_dict_in_model(state_dict)
-        else:
-            logger.warning(
-                f"Could not locate the best model at {best_model_path}, if you are running a distributed training "
-                "on multiple nodes, you should activate `--save_on_each_node`."
-            )
-
-    def _get_train_sampler(self) -> Optional[paddle.io.Sampler]:
-        if self.train_dataset is None or not has_length(self.train_dataset):
-            return None
-
-        if self.args.world_size <= 1:
-            return paddle.io.BatchSampler(
-                dataset=self.train_dataset,
-                shuffle=True,
-                batch_size=self.args.per_device_train_batch_size,
-                drop_last=self.args.dataloader_drop_last,
-            )
-
-        return DistributedBatchSampler(
-            self.train_dataset,
-            batch_size=self.args.per_device_train_batch_size,
-            shuffle=True,
-            num_replicas=self.args.dataset_world_size,
-            rank=self.args.dataset_rank,
-            drop_last=self.args.dataloader_drop_last,
-        )
-
-    def _set_state_dict_in_model(self, state_dict):
-        # TODO  @ZHUI paddle need return the results of set_state_dict.
-        logger.info(f"set state-dict :{self.model.set_state_dict(state_dict)}")
-
-    def _print_timer(self):
-        """print timer and clear states"""
-        paddle_timer_info = ""
-        try:
-            from paddle.distributed.fleet.utils.timer_helper import (
-                get_timers as paddle_get_timers,
-            )
-
-            paddle_pipeline_timers = paddle_get_timers()
-            for name, timer in paddle_pipeline_timers.timers.items():
-                elapsed_time = timer.elapsed(reset=False) * 1000.0
-                paddle_timer_info += f" | {name}: {elapsed_time:.2f}"
-            paddle_pipeline_timers.log(paddle_pipeline_timers.timers.keys(), reset=True)
-        except ImportError:  # paddle version too old, timer not support
-            warnings.warn(f"paddle version:{paddle.__git_commit__} does not support pipeline timer")
-        except AssertionError:  # paddle timer not enabled
-            pass
-
-        if self.timers is not None:
-            timer_info = self.timers.log(self.timers.timers.keys(), reset=True)
-        else:
-            timer_info = ""
-
-        if timer_info or paddle_timer_info:
-            logger.info(f"[Profile global_step: {self.state.global_step}] {timer_info} {paddle_timer_info}")
-
-    def _maybe_log_save_evaluate(self, tr_loss, model, epoch, ignore_keys_for_eval, **kwargs):
-        if self.control.should_log:
-
-            logs: Dict[str, float] = {}
-
-            # all_gather + mean() to get average loss over all processes
-            tr_loss_scalar = self._nested_gather(tr_loss).mean().item()
-
-            # reset tr_loss to zero
-            tr_loss.subtract_(tr_loss)
-
-            logs["loss"] = round(tr_loss_scalar / (self.state.global_step - self._globalstep_last_logged), 8)
-            logs["learning_rate"] = float("{0:.3e}".format(self._get_learning_rate()))
-            logs["global_step"] = int(self.state.global_step)
-
-            total_train_batch_size = (
-                self.args.train_batch_size * self.args.gradient_accumulation_steps * self.args.dataset_world_size
-            )
-            num_steps = self.state.global_step - self._globalstep_last_logged
-            logs.update(
-                speed_metrics(
-                    "interval",
-                    self._globalstep_last_start_time,
-                    num_samples=total_train_batch_size * num_steps,
-                    num_steps=num_steps,
-                )
-            )
-
-            self._total_loss_scalar += tr_loss_scalar
-            self._globalstep_last_logged = self.state.global_step
-            self._globalstep_last_start_time = time.time()
-
-            self.log(logs, **kwargs)
-
-        metrics = None
-        if self.control.should_evaluate:
-            if isinstance(self.optimizer, GroupShardedOptimizerStage2) and self.optimizer._broadcast_overlap:
-                paddle.device.cuda.synchronize()
-
-            if isinstance(self.eval_dataset, dict):
-                for eval_dataset_name, eval_dataset in self.eval_dataset.items():
-                    metrics = self.evaluate(
-                        eval_dataset=eval_dataset,
-                        ignore_keys=ignore_keys_for_eval,
-                        metric_key_prefix=f"eval_{eval_dataset_name}",
-                    )
-            else:
-                metrics = self.evaluate(ignore_keys=ignore_keys_for_eval)
-
-        if self.control.should_save:
-            if isinstance(self.optimizer, GroupShardedOptimizerStage2) and self.optimizer._broadcast_overlap:
-                paddle.device.cuda.synchronize()
-
-            self._save_checkpoint(model, metrics=metrics)
-            self.control = self.callback_handler.on_save(self.args, self.state, self.control)
-
-    def _get_learning_rate(self):
-        return self.optimizer.get_lr()
-
-    def get_train_dataloader(self):
-        """
-        Returns the training [`~paddle.io.DataLoader`].
-
-        Will use no sampler if `self.train_dataset` does not implement `__len__`, a random sampler (adapted to
-        distributed training if necessary) otherwise.
-
-        Subclass and override this method if you want to inject some custom behavior.
-        """
-        if self.args.should_load_dataset and self.train_dataset is None:
-            raise ValueError("Training requires a train_dataset when should_load_dataset is True.")
-        if not self.args.should_load_dataset and self.train_dataset is not None:
-            raise ValueError("We don't need train_dataset when should_load_dataset is False.")
-
-        train_dataset = self.train_dataset
-        if is_datasets_available() and train_dataset is not None and isinstance(train_dataset, datasets.Dataset):
-            train_dataset = self._remove_unused_columns(train_dataset, description="training")
-        _DataLoader = DistDataLoader if self.args.distributed_dataloader else DataLoader
-
-        if self._is_iterable_dataset(train_dataset):
-            if self.args.dataset_world_size > 1:
-                train_dataset = IterableDatasetShard(
-                    train_dataset,
-                    batch_size=self.args.per_device_train_batch_size,
-                    drop_last=self.args.dataloader_drop_last,
-                    num_processes=self.args.dataset_world_size,
-                    process_index=self.args.dataset_rank,
-                )
-
-            return _DataLoader(
-                train_dataset,
-                batch_size=self.args.per_device_train_batch_size,
-                collate_fn=self.data_collator,
-                num_workers=self.args.dataloader_num_workers,
-            )
-
-        if self.args.should_load_dataset:
-            train_sampler = self._get_train_sampler()
-        else:
-            train_sampler = None
-
-        if self.args.distributed_dataloader:
-            logger.info("Training using DistDataLoader.")
-
-        return _DataLoader(
-            train_dataset,
-            batch_sampler=train_sampler,
-            collate_fn=self.data_collator,
-            num_workers=self.args.dataloader_num_workers,
-        )
-
-    def create_optimizer_and_scheduler(self, num_training_steps: int):
-        """
-        Setup the optimizer and the learning rate scheduler.
-
-        We provide a reasonable default that works well. If you want to use something else, you can pass a tuple in the
-        Trainer's init through `optimizers`, or subclass and override this method (or `create_optimizer` and/or
-        `create_scheduler`) in a subclass.
-        """
-        self.create_scheduler(num_training_steps=num_training_steps)
-        self.create_optimizer(self.lr_scheduler)
-
-    def create_optimizer(self, lr_scheduler=None):
-        """
-        Setup the optimizer.
-
-        We provide a reasonable default that works well. If you want to use something else, you can pass a tuple in the
-        Trainer's init through `optimizers`, or subclass and override this method in a subclass.
-        """
-        if self.optimizer is None:
-            if self.optimizer_grouped_parameters is not None:
-                params = self.optimizer_grouped_parameters
-                apply_decay_param_fun = None
-            else:
-                params = self.model.parameters()
-                decay_parameters = [
-                    p.name for n, p in self.model.named_parameters() if not any(nd in n for nd in ["bias", "norm"])
-                ]
-
-                def apply_decay_param_fun(x):
-                    return x in decay_parameters
-
-            optimizer_cls, optimizer_kwargs = Trainer.get_optimizer_cls_and_kwargs(self.args)
-            if hasattr(optimizer_cls, "_create_master_weight") and self.args.fp16_opt_level == "O2":
-                optimizer_kwargs["multi_precision"] = True
-
-            def is_new_version_sharding_stage1_optimizer():
-                signature_keys = set(inspect.signature(DygraphShardingOptimizer).parameters.keys())
-                return "inner_optimizer_class" not in signature_keys
-
-            if ShardingOption.SHARD_OP in self.args.sharding and not is_new_version_sharding_stage1_optimizer():
-                # for backward compatibility.
-                # this call will raise, if sharding stage1 is supported in HybridParallelOptimizer,
-                # in which case, the logic follows will handle it
-                self.optimizer = DygraphShardingOptimizer(
-                    hcg=fleet.get_hybrid_communicate_group(),
-                    user_defined_strategy=None,
-                    params=params,
-                    inner_optimizer_class=optimizer_cls,
-                    learning_rate=self.lr_scheduler if lr_scheduler is None else lr_scheduler,
-                    apply_decay_param_fun=apply_decay_param_fun,
-                    weight_decay=self.args.weight_decay,
-                    grad_clip=nn.ClipGradByGlobalNorm(self.args.max_grad_norm)
-                    if self.args.max_grad_norm > 0
-                    else None,
-                    **optimizer_kwargs,
-                )
-            else:
-                self.optimizer = optimizer_cls(
-                    learning_rate=self.lr_scheduler if lr_scheduler is None else lr_scheduler,
-                    apply_decay_param_fun=apply_decay_param_fun,
-                    parameters=params,
-                    weight_decay=self.args.weight_decay,
-                    grad_clip=nn.ClipGradByGlobalNorm(self.args.max_grad_norm)
-                    if self.args.max_grad_norm > 0
-                    else None,
-                    **optimizer_kwargs,
-                )
-
-        return self.optimizer
-
-    def _load_rng_state(self, checkpoint):
-        # Load RNG states from `checkpoint`
-        if checkpoint is None:
-            return
-
-        # if use distributed training
-        if self.args.world_size > 1:
-            process_index = self.args.process_index
-            rng_file_list = [None for x in range(self.args.world_size)]
-            if self.args.should_save:
-                rng_file = os.path.join(checkpoint, f"rng_state_{self.args.world_size}.pth")
-                if os.path.isfile(rng_file):
-                    rng_file_list = paddle.load(rng_file, return_numpy=True)
-            paddle.distributed.broadcast_object_list(rng_file_list, src=0)
-            # if rng_file_list still empty, then use old style rng_state
-            if rng_file_list[0] is None:
-                rng_file = os.path.join(checkpoint, f"rng_state_{process_index}.pth")
-                if not os.path.isfile(rng_file):
-                    logger.info(
-                        f"Didn't find an RNG file for process {process_index}, if you are resuming a training that "
-                        "wasn't launched in a distributed fashion, reproducibility is not guaranteed."
-                    )
-                    return
-                checkpoint_rng_state = paddle.load(rng_file, return_numpy=True)
-            else:
-                checkpoint_rng_state = rng_file_list[process_index]
-        else:
-            rng_file = os.path.join(checkpoint, "rng_state.pth")
-            if not os.path.isfile(rng_file):
-                logger.info(
-                    "Didn't find an RNG file, if you are resuming a training that was launched in a distributed "
-                    "fashion, reproducibility is not guaranteed."
-                )
-                return
-
-            checkpoint_rng_state = paddle.load(rng_file, return_numpy=True)
-
-        random.setstate(checkpoint_rng_state["python"])
-        np.random.set_state(checkpoint_rng_state["numpy"])
-
-        core = paddle.framework.core
-
-        core.default_cpu_generator().manual_seed(checkpoint_rng_state["cpu"])
-        if core.is_compiled_with_cuda():
-            if not len(checkpoint_rng_state["cuda"]) == core.get_cuda_device_count():
-                raise ValueError("Length of gpu state list shoule be equal to the gpu device count")
-            for i in range(core.get_cuda_device_count()):
-                core.default_cuda_generator(i).manual_seed(checkpoint_rng_state["cuda"][i])
-
-        if paddle.device.get_all_custom_device_type() is not None:
-            custom_device_type = paddle.device.get_all_custom_device_type()
-            for device in custom_device_type:
-                if not len(checkpoint_rng_state["cuda"]) == core.get_custom_device_count(device):
-                    raise ValueError("Length of custom device state list shoule be equal to the custom device count")
-                for i in range(core.get_custom_device_count(device)):
-                    core.default_custom_device_generator(i).manual_seed(checkpoint_rng_state["cuda"][i])
-
-        if self.args.use_hybrid_parallel:
-            fleet.meta_parallel.get_rng_state_tracker().set_states_tracker(
-                checkpoint_rng_state["hybrid_parallel_rng_state_tracker"]
-            )
-
-    @staticmethod
-    def get_optimizer_cls_and_kwargs(args: TrainingArguments) -> Tuple[Any, Any]:
-        """
-        Returns the optimizer class and optimizer parameters based on the training arguments.
-
-        Args:
-            args (`paddlenlp.training_args.TrainingArguments`):
-                The training arguments for the training session.
-
-        """
-        # optimizer_kwargs = {"lr": args.learning_rate}
-        optimizer_kwargs = {}
-        adam_kwargs = {
-            "beta1": args.adam_beta1,
-            "beta2": args.adam_beta2,
-            "epsilon": args.adam_epsilon,
-        }
-        if args.optim == OptimizerNames.ADAMW:
-            from paddle.optimizer import AdamW
-
-            optimizer_cls = AdamW
-            optimizer_kwargs.update(adam_kwargs)
-        else:
-            raise ValueError(f"Trainer cannot instantiate unsupported optimizer: {args.optim}")
-        return optimizer_cls, optimizer_kwargs
-
-    def create_scheduler(self, num_training_steps: int):
-        """
-        Setup the scheduler. The optimizer of the trainer must have been set up either before this method is called or
-        passed as an argument.
-
-        Args:
-            num_training_steps (int): The number of training steps to do.
-        """
-        warmup = (
-            self.args.warmup_steps if self.args.warmup_steps > 0 else int(self.args.warmup_ratio * num_training_steps)
-        )
-
-        if self.lr_scheduler is None:
-            self.lr_scheduler = get_scheduler(
-                self.args.lr_scheduler_type,
-                learning_rate=self.args.learning_rate,
-                num_warmup_steps=warmup,
-                num_training_steps=num_training_steps,
-                num_cycles=self.args.num_cycles,
-                lr_end=self.args.lr_end,
-                power=self.args.power,
-            )
-
-        return self.lr_scheduler
-
-    def num_examples(self, dataloader: DataLoader) -> int:
-        """
-        Helper to get number of samples in a [`~paddle.io.DataLoader`] by accessing its dataset. When
-        dataloader.dataset does not exist or has no length, estimates as best it can
-        """
-        try:
-            dataset = dataloader.dataset
-            # Special case for IterableDatasetShard, we need to dig deeper
-            if isinstance(dataset, IterableDatasetShard):
-                return len(dataloader.dataset.dataset)
-            return len(dataloader.dataset)
-        except (NameError, AttributeError, TypeError):  # no dataset or length, estimate by length of dataloader
-            return len(dataloader) * self.args.per_device_train_batch_size
-
-    def _wrap_model(self, model, training=True):
-
-        # train/eval could be run multiple-times - if already wrapped, don't re-wrap it again
-        if unwrap_model(model) is not model:
-            return model
-
-        # Note: in paddle.distributed mode, there's no point in wrapping the model
-        # inside a DistributedDataParallel as we'll be under `no_grad` anyways.
-        if not training:
-            return model
-
-        # Mixed precision training
-        if training and self.do_grad_scaling:  # self.args.fp16_opt_level=="O2":
-            # model, self.optimizer
-            if self.amp_dtype == "bfloat16":
-                # fix for paddlepaddle < 2.4.1, not support for bf16
-                decorated = paddle.amp.decorate(
-                    models=model, optimizers=self.optimizer, level=self.args.fp16_opt_level, dtype=self.amp_dtype
-                )
-            else:
-                decorated = paddle.amp.decorate(
-                    models=model, optimizers=self.optimizer, level=self.args.fp16_opt_level
-                )
-
-            if self.optimizer is None:
-                model = decorated
-            else:
-                model, self.optimizer = decorated
-
-        # Multi-gpu training
-        if self.args.world_size > 1 and not self.args.use_hybrid_parallel:
-            model = paddle.DataParallel(model)
-            # Distributed training (should be after fp16 initialization)
-
-        in_pipeline_parallel_mode = self.args.pipeline_parallel_degree > 1
-        in_sharding_parallel_mode = self.sharding is not None
-        in_tensor_parallel_model = self.args.tensor_parallel_degree > 1
-
-        # Pipeline mode
-        if in_pipeline_parallel_mode:
-            if self.args.amp_master_grad:
-                mix_precision_utils.MixPrecisionLayer(model, dtype=self.amp_dtype)  # return value has no use
-            # hack for pipeline model mini batch to batch
-            # need batter solution @ZHUI
-            # make batch_fn compatible for fleet.distributed_model decorate.
-            prepare_pipeline_inputs_func = (
-                model._prepare_pipeline_inputs_func if hasattr(model, "_prepare_pipeline_inputs_func") else None
-            )
-            if isinstance(model, LoRAModel):
-                model = model.model
-            model = fleet.distributed_model(model)
-            if prepare_pipeline_inputs_func is not None:
-                model._prepare_pipeline_inputs_func = prepare_pipeline_inputs_func
-            else:
-
-                def _prepare_pipeline_inputs_func(inputs):
-                    first_stage_keys = ["input_ids", "attention_mask", "position_ids"]
-                    last_stage_keys = ["labels"]
-
-                    def get_expected_keys(inputs, keys):
-                        ret = tuple([inputs.pop(k) for k in keys if k in inputs])
-                        if len(ret) == 1:
-                            ret = ret[0]
-                        return ret
-
-                    if type(inputs) is dict or type(inputs) is OrderedDict:
-                        return [
-                            get_expected_keys(inputs, first_stage_keys),
-                            get_expected_keys(inputs, last_stage_keys),
-                        ]
-
-                    keys = list(inputs[0].keys())
-                    inputs_batch = {key: [data.pop(key) for data in inputs] for key in keys}
-                    return [
-                        get_expected_keys(inputs_batch, first_stage_keys),
-                        get_expected_keys(inputs_batch, last_stage_keys),
-                    ]
-
-                logger.warning(
-                    "Using default prepare pipeline inputs func, only support input_ids and labels as inputs."
-                )
-                model._prepare_pipeline_inputs_func = _prepare_pipeline_inputs_func
-
-            assert self.optimizer is not None, "Pipeline mode need decorate optimizer, pelease init optimizer."
-            if self.args.amp_master_grad:
-                self.optimizer = mix_precision_utils.MixPrecisionOptimizer(self.optimizer)
-            self.optimizer = fleet.distributed_optimizer(self.optimizer)
-
-        # No pipeline mode, sharding only
-        if not in_pipeline_parallel_mode and in_sharding_parallel_mode:
-            # Sharded DDP!
-            if self.args.tensor_parallel_degree > 1:
-                hcg = fleet.get_hybrid_communicate_group()
-                assert (
-                    ShardingOption.SHARD_GRAD_OP in self.args.sharding or ShardingOption.SHARD_OP in self.args.sharding
-                ), "Only support tensor parallel + sharding stage1/stage2 hybrid parallel now."
-                model = paddle.distributed.fleet.meta_parallel.TensorParallel(model, hcg, strategy=None)
-
-            if ShardingOption.SHARD_OP in self.args.sharding:
-                if self.args.amp_master_grad:
-                    mix_precision_utils.MixPrecisionLayer(model, dtype=self.amp_dtype)  # return value has no use
-                model = fleet.distributed_model(model)
-                if self.args.amp_master_grad:
-                    self.optimizer = mix_precision_utils.MixPrecisionOptimizer(self.optimizer)
-                self.optimizer = fleet.distributed_optimizer(self.optimizer)
-            else:
-                # sync params (broadcast) buffers in dp group
-                if not is_dp_group_support_in_group_sharded_parallel() and self.args.data_parallel_degree > 1:
-                    from paddle.distributed.parallel import sync_params_buffers
-
-                    hcg = fleet.get_hybrid_communicate_group()
-                    dp_group = hcg.get_data_parallel_group()
-                    sync_params_buffers(model, comm_group=dp_group, src_rank=dp_group.ranks[0])
-
-                cpu_offload = ShardingOption.OFFLOAD in self.args.sharding
-                assert self.optimizer is not None, "optimizer is empty!"
-                level = None
-                if ShardingOption.SHARD_GRAD_OP in self.args.sharding:
-                    level = "os_g"
-                if ShardingOption.FULL_SHARD in self.args.sharding:
-                    level = "p_g_os"
-
-                from paddle.distributed.sharding import group_sharded_parallel
-
-                # add dp_group and exclude_layer params
-                # https://www.paddlepaddle.org.cn/documentation/docs/zh/develop/api/paddle/distributed/sharding/group_sharded_parallel_cn.html#group-sharded-parallel
-                extra_kwargs = {}
-                if is_dp_group_support_in_group_sharded_parallel():
-                    extra_kwargs["dp_group"] = self.dp_group
-                    extra_kwargs["exclude_layer"] = ["GroupNorm"]
-
-                if self.args.amp_master_grad:
-                    assert level == "os_g", "Main grad doesn't support sharding stage 3 for now."
-                    assert (
-                        self.args.data_parallel_degree == 1
-                    ), "Sharding stage 2 main grad is not compatible with dp for now."
-                    mix_precision_utils.MixPrecisionLayer(model, dtype=self.amp_dtype)  # return value has no use
-                    self.optimizer = mix_precision_utils.MixPrecisionOptimizer(self.optimizer)
-
-                model, optimizer, _ = group_sharded_parallel(
-                    model,
-                    self.optimizer,
-                    level=level,
-                    scaler=None,
-                    group=self.sharding_group,
-                    offload=cpu_offload,
-                    **extra_kwargs,
-                )
-                if self.args.amp_master_grad:
-                    assert hasattr(optimizer, "use_main_grad"), (
-                        "Current installed paddle doesn't support sharding stage 2 with main grad, "
-                        "please upgrade your paddle (using nightly version)."
-                    )
-
-                sharding_parallel_config = set(self.args.sharding_parallel_config.split(" "))
-                if level == "os_g" and "enable_stage2_overlap" in sharding_parallel_config:
-                    model._set_reduce_overlap(True)
-                    optimizer._set_broadcast_overlap(True, model)
-
-                self.optimizer = optimizer
-
-        # pure tesnor parallel mode, no pipeline_parallel, no sharding.
-        if not in_pipeline_parallel_mode and not in_sharding_parallel_mode and in_tensor_parallel_model:
-            if self.args.amp_master_grad:
-                mix_precision_utils.MixPrecisionLayer(model, dtype=self.amp_dtype)  # return value has no use
-
-            model = fleet.distributed_model(model)
-            assert self.optimizer is not None, "Tensor parallel mode need decorate optimizer, pelease init optimizer."
-            if self.args.amp_master_grad:
-                self.optimizer = mix_precision_utils.MixPrecisionOptimizer(self.optimizer)
-            self.optimizer = fleet.distributed_optimizer(self.optimizer)
-
-        return model
-
-    def _prepare_input(self, data: Union[paddle.Tensor, Any]) -> Union[paddle.Tensor, Any]:
-        """
-        Prepares one `data` before feeding it to the model, be it a tensor or a nested list/dictionary of tensors.
-        """
-        if isinstance(data, Mapping):
-            return type(data)({k: self._prepare_input(v) for k, v in data.items()})
-        elif isinstance(data, (tuple, list)):
-            return type(data)(self._prepare_input(v) for v in data)
-        elif isinstance(data, paddle.Tensor):
-            # kwargs = dict(device=self.args.current_device)
-            # update data type for pure fp16
-            if data.place.is_cuda_pinned_place():
-                return data.cuda()
-            return data
-            # return data.to(**kwargs)
-        return data
-
-    def _prepare_inputs(self, inputs: Dict[str, Union[paddle.Tensor, Any]]) -> Dict[str, Union[paddle.Tensor, Any]]:
-        """
-        Prepare `inputs` before feeding them to the model, converting them to tensors if they are not already and
-        handling potential state.
-        """
-        inputs = self._prepare_input(inputs)
-        if self.args.past_index >= 0 and self._past is not None:
-            inputs["mems"] = self._past
-
-        return inputs
-
-    def autocast_smart_context_manager(self):
-        """
-        A helper wrapper that creates an appropriate context manager for `autocast` while feeding it the desired
-        arguments, depending on the situation.
-        """
-        if self.enable_autocast_context_manager:
-            custom_black_list = ["reduce_sum", "c_softmax_with_cross_entropy"]
-            custom_white_list = []
-            if self.args.fp16_opt_level == "O2":
-                # https://github.com/PaddlePaddle/Paddle/blob/eb97f4f0adca40b16a309b927e480178beb8ae96/python/paddle/amp/amp_lists.py#L85-L86
-                # the lookup_table is in black_list, but in O2, we need it return fp16
-                custom_white_list.extend(["lookup_table", "lookup_table_v2"])
-
-            if self.args.amp_custom_white_list is not None:
-                custom_white_list.extend(self.args.amp_custom_white_list)
-            if self.args.amp_custom_black_list is not None:
-                custom_black_list.extend(self.args.amp_custom_black_list)
-
-            ctx_manager = autocast(
-                True,
-                custom_black_list=set(custom_black_list),
-                custom_white_list=set(custom_white_list),
-                level=self.args.fp16_opt_level,
-                dtype=self.amp_dtype,
-            )
-        else:
-            ctx_manager = contextlib.nullcontext() if sys.version_info >= (3, 7) else contextlib.suppress()
-
-        return ctx_manager
-
-    def compute_loss(self, model, inputs, return_outputs=False):
-        """
-        How the loss is computed by Trainer. By default, all models return the loss in the first element.
-        Subclass and override for custom behavior.
-        """
-        if self.criterion is not None:
-            if "labels" in inputs:
-                labels = inputs.pop("labels")
-            elif "start_positions" in inputs and "end_positions" in inputs:
-                labels = (inputs.pop("start_positions"), inputs.pop("end_positions"))
-            elif self.args.label_names is not None:
-                labels = []
-                for label in self.label_names:
-                    labels.append(inputs.pop(label))
-                labels = tuple(labels)
-            elif "generator_labels" in inputs:
-                labels = inputs["generator_labels"]
-        else:
-            labels = None
-
-        outputs = model(**inputs)
-
-        if self.criterion is not None:
-            loss = self.criterion(outputs, labels)
-            outputs = (loss, outputs)
-
-        # Save past state if it exists
-        # TODO: this needs to be fixed and made cleaner later.
-        if self.args.past_index >= 0:
-            self._past = outputs[self.args.past_index]
-
-        # We don't use .loss here since the model may return tuples instead of ModelOutput.
-        loss = outputs["loss"] if isinstance(outputs, dict) else outputs
-        if isinstance(outputs, dict):
-            loss = outputs["loss"]
-        elif isinstance(outputs, tuple):
-            loss = outputs[0]
-        else:
-            loss = outputs
-
-        return (loss, outputs) if return_outputs else loss
-
-    def training_step(self, model: nn.Layer, inputs: Dict[str, Union[paddle.Tensor, Any]]) -> paddle.Tensor:
-        """
-        Perform a training step on a batch of inputs.
-
-        Subclass and override to inject custom behavior.
-
-        Args:
-            model (`nn.Layer`):
-                The model to train.
-            inputs (`Dict[str, Union[paddle.Tensor, Any]]`):
-                The inputs and targets of the model.
-
-                The dictionary will be unpacked before being fed to the model. Most models expect the targets under the
-                argument `labels`. Check your model's documentation for all accepted arguments.
-
-        Return:
-            `paddle.Tensor`: The tensor with training loss on this batch.
-        """
-        if self.args.pipeline_parallel_degree > 1:
-            return self.training_pipeline_step(model, inputs)
-
-        model.train()
-        inputs = self._prepare_inputs(inputs)
-
-        with self.autocast_smart_context_manager():
-            loss = self.compute_loss(model, inputs)
-
-        if self.args.gradient_accumulation_steps > 1:
-            loss = loss / self.args.gradient_accumulation_steps
-
-        if self.do_grad_scaling:
-            self.scaler.scale(loss).backward()
-        else:
-            loss.backward()
-
-        return loss.detach()
-
-    def training_pipeline_step(self, model: nn.Layer, inputs: Dict[str, Union[paddle.Tensor, Any]]) -> paddle.Tensor:
-        """
-        Perform a training step on a batch of inputs.
-
-        Subclass and override to inject custom behavior.
-
-        Args:
-            model (`nn.Layer`):
-                The model to train.
-            inputs (`Dict[str, Union[paddle.Tensor, Any]]`):
-                The inputs and targets of the model.
-
-                The dictionary will be unpacked before being fed to the model. Most models expect the targets under the
-                argument `labels`. Check your model's documentation for all accepted arguments.
-
-        Return:
-            `paddle.Tensor`: The tensor with training loss on this batch.
-        """
-        # accumulation data
-        if not hasattr(self, "_pp_data_buffer"):
-            self._pp_data_buffer = []
-        self._pp_data_buffer.append(inputs)
-        if len(self._pp_data_buffer) != self.args.gradient_accumulation_steps:
-            return paddle.zeros([])
-
-        # for v in self._pp_data_buffer[0].values():
-        #     assert isinstance(v, paddle.Tensor), f"Only support tensor as pipeline mode input, got type {type(v)}"
-
-        inputs = model._prepare_pipeline_inputs_func(self._pp_data_buffer)
-        self._pp_data_buffer = []
-
-        model.train()
-        # hack pipeline-layers
-        # since the pipeline layer will check input is valid every iter.
-        # in same case,  for example, batch size warmup, we need dynamic change gradient_accumulation_steps to implement.
-        config_backup = model.micro_batch_size, model.accumulate_steps
-        model.micro_batch_size = self.args.per_device_train_batch_size
-        model.accumulate_steps = self.args.gradient_accumulation_steps
-
-        if model._dp_comm_overlap or model._sharding_comm_overlap:
-            for _, buffers in model._chunk_2_comm_buffers.items():
-                for buffer in buffers:
-                    buffer._acc_steps = self.args.gradient_accumulation_steps
-
-        inputs = model._prepare_training(
-            inputs, self.optimizer, self.lr_scheduler
-        )  # None, None => [optimizer, lr_scheduler]
-        model.optimizer = None  # we do not use `PipelineParallel` to handler optimizer step
-        model.lr_scheduler = None
-
-        with self.autocast_smart_context_manager():
-            loss = model.forward_backward_pipeline(inputs, self.scaler if self.do_grad_scaling else None)
-
-        model.micro_batch_size, model.accumulate_steps = config_backup
-
-        return loss.detach()
-
-    def save_model(self, output_dir: Optional[str] = None, merge_tensor_parallel: Optional[bool] = False):
-        """
-        Will save the model, so you can reload it using `from_pretrained()`.
-
-        Will only save from the main process.
-        """
-
-        if output_dir is None:
-            output_dir = self.args.output_dir
-
-        if ShardingOption.FULL_SHARD in self.args.sharding:
-            self.model_wrapped.get_all_parameters(convert2cpu=True)
-
-        if self.args.should_save_model_state:
-            self._save(output_dir=output_dir, merge_tensor_parallel=merge_tensor_parallel)
-
-    def _save_checkpoint(self, model, metrics=None):
-        # assert unwrap_model(model) is self.model, "internal model should be a reference to self.model"
-
-        # Save model checkpoint
-        checkpoint_folder = f"{PREFIX_CHECKPOINT_DIR}-{self.state.global_step}"
-
-        run_dir = self.args.output_dir
-
-        output_dir = os.path.join(run_dir, checkpoint_folder)
-
-        if isinstance(self.model, LoRAModel) and (self.model.quantized or self.args.pipeline_parallel_degree > 1):
-            self.save_model(output_dir)
-        elif isinstance(self.model, LoRAModel) or isinstance(self.model, PrefixModelForCausalLM):
-            self.save_model(output_dir, True)
-        else:
-            self.save_model(output_dir)
-
-        optimizer_name = _add_variant(OPTIMIZER_NAME, self.args.optimizer_name_suffix)
-
-        if self.args.use_hybrid_parallel:
-            if self.dp_group.rank <= 0:
-                os.makedirs(output_dir, exist_ok=True)
-                logger.info("Saving optimizer files.")
-                paddle.save(
-                    self.optimizer.state_dict(),
-                    os.path.join(output_dir, optimizer_name),
-                )
-
-        if self.args.should_save:
-            if not self.args.use_hybrid_parallel:
-                logger.info("Saving optimizer files.")
-                paddle.save(self.optimizer.state_dict(), os.path.join(output_dir, OPTIMIZER_NAME))
-
-            # FIXME: manybe only save one copy
-            paddle.save(self.lr_scheduler.state_dict(), os.path.join(output_dir, SCHEDULER_NAME))
-
-            if self.do_grad_scaling:
-                paddle.save(self.scaler.state_dict(), os.path.join(output_dir, SCALER_NAME))
-
-        # Determine the new best metric / best model checkpoint
-        if metrics is not None and self.args.metric_for_best_model is not None:
-            metric_to_check = self.args.metric_for_best_model
-            if not metric_to_check.startswith("eval_"):
-                metric_to_check = f"eval_{metric_to_check}"
-            metric_value = metrics[metric_to_check]
-
-            operator = np.greater if self.args.greater_is_better else np.less
-            if (
-                self.state.best_metric is None
-                or self.state.best_model_checkpoint is None
-                or operator(metric_value, self.state.best_metric)
-            ):
-                self.state.best_metric = metric_value
-                self.state.best_model_checkpoint = output_dir
-
-        # Save the Trainer state
-        if self.args.should_save:
-            self.state.save_to_json(os.path.join(output_dir, TRAINER_STATE_NAME))
-
-        # Save RNG state in non-distributed training
-        rng_states = {
-            "python": random.getstate(),
-            "numpy": np.random.get_state(),
-            "cuda": [k.current_seed() for k in paddle.get_rng_state()],
-            "cpu": paddle.framework.core.default_cpu_generator().get_state().current_seed(),
-        }
-        if self.args.use_hybrid_parallel:
-            rng_states[
-                "hybrid_parallel_rng_state_tracker"
-            ] = fleet.meta_parallel.get_rng_state_tracker().get_states_tracker()
-
-        if self.args.world_size > 1:
-            rng_states_list = []
-            paddle.distributed.all_gather_object(rng_states_list, rng_states)
-            if self.args.should_save:
-                os.makedirs(output_dir, exist_ok=True)
-                paddle.save(rng_states_list, os.path.join(output_dir, f"rng_state_{self.args.world_size}.pth"))
-        else:
-            os.makedirs(output_dir, exist_ok=True)
-            paddle.save(rng_states, os.path.join(output_dir, "rng_state.pth"))
-
-        # Maybe delete some older checkpoints.
-        # For hybrid parallel training, the checkpoint files maybe on different node.
-        if self.args.should_save_model_state and self.args.local_rank == 0:
-            self._rotate_checkpoints(use_mtime=True, output_dir=run_dir)
-
-    def set_optimizer_grouped_parameters(self, optimizer_grouped_parameters=None):
-        """
-        set optimizer grouped parameters:
-
-        you can set optimizer_grouped_parameters with whatever argments on whatever parameters to train.
-        """
-        self.optimizer_grouped_parameters = optimizer_grouped_parameters
-
-    def disable_autocast_context_manager(self):
-        """
-        For pure fp16 or pure bf16 training, the paddle.amp.autocast is annoy for always cast fp32 to fp16.
-        if you networks cast fp16 to fp32 manually to get higher precision, autocast make it not work, since it cast fp32 to fp16 back.
-
-        """
-        assert self.args.fp16_opt_level == "O2", "disable_autocast_context_manager should only work for pure fp16/bf16"
-        self.enable_autocast_context_manager = False
-
-    def _sorted_checkpoints(
-        self, output_dir=None, checkpoint_prefix=PREFIX_CHECKPOINT_DIR, use_mtime=False
-    ) -> List[str]:
-        ordering_and_checkpoint_path = []
-
-        glob_checkpoints = [str(x) for x in Path(output_dir).glob(f"{checkpoint_prefix}-*")]
-
-        for path in glob_checkpoints:
-            if use_mtime:
-                ordering_and_checkpoint_path.append((os.path.getmtime(path), path))
-            else:
-                regex_match = re.match(f".*{checkpoint_prefix}-([0-9]+)", path)
-                if regex_match is not None and regex_match.groups() is not None:
-                    ordering_and_checkpoint_path.append((int(regex_match.groups()[0]), path))
-
-        checkpoints_sorted = sorted(ordering_and_checkpoint_path)
-        checkpoints_sorted = [checkpoint[1] for checkpoint in checkpoints_sorted]
-        # Make sure we don't delete the best model.
-        if self.state.best_model_checkpoint is not None:
-            best_model_index = checkpoints_sorted.index(str(Path(self.state.best_model_checkpoint)))
-            for i in range(best_model_index, len(checkpoints_sorted) - 2):
-                checkpoints_sorted[i], checkpoints_sorted[i + 1] = checkpoints_sorted[i + 1], checkpoints_sorted[i]
-        return checkpoints_sorted
-
-    def _rotate_checkpoints(self, use_mtime=False, output_dir=None) -> None:
-        if self.args.save_total_limit is None or self.args.save_total_limit <= 0:
-            return
-
-        # Check if we should delete older checkpoint(s)
-        checkpoints_sorted = self._sorted_checkpoints(use_mtime=use_mtime, output_dir=output_dir)
-        if len(checkpoints_sorted) <= self.args.save_total_limit:
-            return
-
-        # If save_total_limit=1 with load_best_model_at_end=True, we could end up deleting the last checkpoint, which
-        # we don't do to allow resuming.
-        save_total_limit = self.args.save_total_limit
-        if (
-            self.state.best_model_checkpoint is not None
-            and self.args.save_total_limit == 1
-            and checkpoints_sorted[-1] != self.state.best_model_checkpoint
-        ):
-            save_total_limit = 2
-
-        number_of_checkpoints_to_delete = max(0, len(checkpoints_sorted) - save_total_limit)
-        checkpoints_to_be_deleted = checkpoints_sorted[:number_of_checkpoints_to_delete]
-        for checkpoint in checkpoints_to_be_deleted:
-            logger.info(f"Deleting older checkpoint [{checkpoint}] due to args.save_total_limit")
-            shutil.rmtree(checkpoint)
-
-    def _save(self, output_dir: Optional[str] = None, state_dict=None, merge_tensor_parallel=False):
-        output_dir = output_dir if output_dir is not None else self.args.output_dir
-        os.makedirs(output_dir, exist_ok=True)
-        logger.info(f"Saving model checkpoint to {output_dir}")
-        # Save a trained model and configuration using `save_pretrained()`.
-        # They can then be reloaded using `from_pretrained()`
-
-        if self.args.should_save:
-            if self.tokenizer is not None:
-                self.tokenizer.save_pretrained(output_dir)
-            # Good practice: save your training arguments together with the trained model
-            paddle.save(self.args, os.path.join(output_dir, TRAINING_ARGS_NAME))
-
-        if self.args.unified_checkpoint:
-            save_unified_checkpoint(self.args, self.model, output_dir, safe_serialization=True)
-            return
-
-        merge_tensor_parallel = merge_tensor_parallel and self.args.use_hybrid_parallel
-        # peft model
-        if isinstance(self.model, LoRAModel) or isinstance(self.model, PrefixModelForCausalLM):
-            self.model.save_pretrained(
-                output_dir,
-                variant=self.args.weight_name_suffix,
-                merge_tensor_parallel=merge_tensor_parallel,
-                is_main_process=self.args.should_save,
-                max_shard_size="1024GB",
-            )
-        # TODO: @ZHUI unifiy unwrap_model(self.model) and self.model
-        elif not isinstance(self.model, PretrainedModel):
-            if isinstance(unwrap_model(self.model), PretrainedModel):
-                if self.args.should_save_sharding_stage1_model:
-                    config_to_save = None
-                    state_dict, config_to_save, weight_name_suffix = self.sharding_io.manipulate_state_dict_and_config(
-                        unwrap_model(self.model), merge_tensor_parallel=merge_tensor_parallel
-                    )
-                    unwrap_model(self.model).save_pretrained(
-                        output_dir,
-                        state_dict=state_dict,
-                        config_to_save=config_to_save,
-                        merge_tensor_parallel=merge_tensor_parallel,
-                        variant=weight_name_suffix,
-                        is_main_process=self.args.should_save,
-                        max_shard_size="1024GB",
-                    )
-                else:
-                    unwrap_model(self.model).save_pretrained(
-                        output_dir,
-                        merge_tensor_parallel=merge_tensor_parallel,
-                        variant=self.args.weight_name_suffix,
-                        is_main_process=self.args.should_save,
-                        max_shard_size="1024GB",
-                    )
-            else:
-                logger.info("Trainer.model is not a `PretrainedModel`, only saving its state dict.")
-                if merge_tensor_parallel:
-                    logger.warning("Trainer.model is not a `PretrainedModel`, not suppor for merge_tensor_parallel.")
-                if state_dict is None:
-                    state_dict = self.model.state_dict()
-                paddle.save(
-                    state_dict,
-                    os.path.join(output_dir, _add_variant(PADDLE_WEIGHTS_NAME, self.args.weight_name_suffix)),
-                )
-        else:
-            if isinstance(self.model, PretrainedModel) and self.args.should_save_sharding_stage1_model:
-                config_to_save = None
-                state_dict, config_to_save, weight_name_suffix = self.sharding_io.manipulate_state_dict_and_config(
-                    self.model, merge_tensor_parallel=merge_tensor_parallel
-                )
-                self.model.save_pretrained(
-                    output_dir,
-                    state_dict=state_dict,
-                    config_to_save=config_to_save,
-                    merge_tensor_parallel=merge_tensor_parallel,
-                    variant=weight_name_suffix,
-                    is_main_process=self.args.should_save,
-                    max_shard_size="1024GB",
-                )
-            else:
-                self.model.save_pretrained(
-                    output_dir,
-                    merge_tensor_parallel=merge_tensor_parallel,
-                    variant=self.args.weight_name_suffix,
-                    is_main_process=self.args.should_save,
-                    max_shard_size="1024GB",
-                )
-        if self.args.should_save_sharding_stage1_model:
-            self.sharding_io.save_distributed_model_meta(output_dir)
-
-    def _load_optimizer_and_scheduler(self, checkpoint):
-        """If optimizer and scheduler states exist, load them."""
-        # TODO: Support DP broadcast optimizer.
-        if checkpoint is None:
-            return
-
-        opt_state_dict = None
-        if self.args.should_load_sharding_stage1_model:
-            opt_state_dict = self.sharding_io.load_optimizer_state_with_reshard(
-                checkpoint, OPTIMIZER_NAME, self.model_wrapped
-            )
-        else:
-            optimizer_name = _add_variant(OPTIMIZER_NAME, self.args.optimizer_name_suffix)
-            if self.args.data_parallel_rank == 0:
-                path = os.path.join(checkpoint, optimizer_name)
-                if os.path.isfile(path):
-                    opt_state_dict = paddle.load(path)
-
-        # broadcast optimizer state in dp group
-        opt_state_dict = broadcast_dp_optimizer(opt_state_dict)
-
-        if opt_state_dict is not None:
-            # Load in optimizer and scheduler states
-            self.optimizer.set_state_dict(opt_state_dict)
-        else:
-            raise ValueError(f"optimizer-state-dict not found, opt: {os.path.join(checkpoint, optimizer_name)}.")
-
-        if distributed_isfile(os.path.join(checkpoint, SCHEDULER_NAME)):
-            self.lr_scheduler.set_state_dict(paddle.load(distributed_file(os.path.join(checkpoint, SCHEDULER_NAME))))
-        else:
-            raise ValueError(f"scheduler-file not found, scheduler:{os.path.join(checkpoint, SCHEDULER_NAME)}")
-
-        if self.do_grad_scaling and distributed_isfile(os.path.join(checkpoint, SCALER_NAME)):
-            self.scaler.load_state_dict(
-                paddle.load(distributed_file(os.path.join(checkpoint, SCALER_NAME)), return_numpy=True)
-            )
-
-    def log(self, logs: Dict[str, float], **kwargs) -> None:
-        """
-        Log `logs` on the various objects watching training.
-
-        Subclass and override this method to inject custom behavior.
-
-        Args:
-            logs (`Dict[str, float]`):
-                The values to log.
-        """
-
-        try:
-            from paddle.distributed.fleet.utils.timer_helper import (
-                get_timers as paddle_get_timers,
-            )
-
-            paddle_pipeline_timers = paddle_get_timers()
-        except ImportError:  # paddle version too old, timer not support
-            warnings.warn(f"paddle version:{paddle.__git_commit__} does not support pipeline timer")
-            paddle_pipeline_timers = None
-        except AssertionError:
-            paddle_pipeline_timers = None
-        kwargs.update(timer=self.timers, paddle_pipeline_timers=paddle_pipeline_timers)
-
-        if self.state.epoch is not None:
-            logs["epoch"] = round(self.state.epoch, 4)
-        output = {**logs, **{"step": self.state.global_step}}
-        self.state.log_history.append(output)
-        self.control = self.callback_handler.on_log(self.args, self.state, self.control, logs, **kwargs)
-
-    def is_local_process_zero(self) -> bool:
-        """
-        Whether or not this process is the local (e.g., on one machine if training in a distributed fashion on several
-        machines) main process.
-        """
-        return self.args.local_process_index == 0
-
-    def is_world_process_zero(self) -> bool:
-        """
-        Whether or not this process is the global main process (when training in a distributed fashion on several
-        machines, this is only going to be `True` for one process).
-        """
-        return self.args.process_index == 0
-
-    def _nested_gather(self, tensors):
-        """
-        Gather value of `tensors` (tensor or list/tuple of nested tensors) and convert them to numpy before
-        concatenating them to `gathered`
-        """
-        if tensors is None:
-            return
-        if self.args.local_rank != -1:
-            tensors = distributed_concat(tensors)
-        return tensors
-
-        # Copied from Accelerate.
-
-    def _pad_across_processes(self, tensor, pad_index=-100):
-        """
-        Recursively pad the tensors in a nested list/tuple/dictionary of tensors from all devices to the same size so
-        they can safely be gathered.
-        """
-        if isinstance(tensor, (list, tuple)):
-            return type(tensor)(self._pad_across_processes(t, pad_index=pad_index) for t in tensor)
-        elif isinstance(tensor, dict):
-            return type(tensor)({k: self._pad_across_processes(v, pad_index=pad_index) for k, v in tensor.items()})
-        elif not isinstance(tensor, paddle.Tensor):
-            raise TypeError(
-                f"Can't pad the values of type {type(tensor)}, only of nested list/tuple/dicts of tensors."
-            )
-
-        if len(tensor.shape) < 2:
-            return tensor
-        # Gather all sizes
-        size = paddle.to_tensor(tensor.shape)[None]
-        sizes = self._nested_gather(size).cpu()
-
-        max_size = max(s[1] for s in sizes)
-        if tensor.shape[1] == max_size:
-            return tensor
-
-        # Then pad to the maximum size
-        old_size = tensor.shape
-        new_size = list(old_size)
-        new_size[1] = max_size
-        # new_tensor = tensor.new_zeros(tuple(new_size)) + pad_index
-        new_tensor = paddle.zeros(tuple(new_size), dtype=tensor.dtype) + pad_index
-        new_tensor[:, : old_size[1]] = tensor
-        return new_tensor
-
-    def _set_signature_columns_if_needed(self):
-        if self._signature_columns is None:
-            # Inspect model forward signature to keep only the arguments it accepts.
-            signature = inspect.signature(self.model.forward)
-            self._signature_columns = list(signature.parameters.keys())
-            # Labels may be named label or label_ids, the default data collator handles that.
-            self._signature_columns += list(set(["label", "label_ids"] + self.label_names))
-
-    def _remove_unused_columns(self, dataset: "datasets.Dataset", description: Optional[str] = None):
-        if not self.args.remove_unused_columns:
-            return dataset
-        if self._signature_columns is None:
-            # Inspect model forward signature to keep only the arguments it accepts.
-            signature = inspect.signature(self.model.forward)
-            self._signature_columns = list(signature.parameters.keys())
-            # Labels may be named label or label_ids, the default data collator handles that.
-            self._signature_columns += ["label", "label_ids", "labels", "start_positions", "end_positions"]
-
-        ignored_columns = list(set(dataset.column_names) - set(self._signature_columns))
-        if len(ignored_columns) > 0:
-            dset_description = "" if description is None else f"in the {description} set "
-            logger.info(
-                f"The following columns {dset_description} don't have a corresponding argument in "
-                f"`{self.model.__class__.__name__}.forward` and have been ignored: {', '.join(ignored_columns)}."
-                f" If {', '.join(ignored_columns)} are not expected by `{self.model.__class__.__name__}.forward`, "
-                f" you can safely ignore this message."
-            )
-
-        columns = [k for k in self._signature_columns if k in dataset.column_names]
-
-        if version.parse(datasets.__version__) < version.parse("1.4.0"):
-            dataset.set_format(
-                type=dataset.format["type"], columns=columns, format_kwargs=dataset.format["format_kwargs"]
-            )
-            return dataset
-        else:
-            return dataset.remove_columns(ignored_columns)
-
-    def _get_collator_with_removed_columns(
-        self, data_collator: Callable, description: Optional[str] = None
-    ) -> Callable:
-        """Wrap the data collator in a callable removing unused columns."""
-        if not self.args.remove_unused_columns:
-            return data_collator
-        self._set_signature_columns_if_needed()
-        signature_columns = self._signature_columns
-
-        remove_columns_collator = RemoveColumnsCollator(
-            data_collator=data_collator,
-            signature_columns=signature_columns,
-            logger=logger,
-            description=description,
-            model_name=self.model.__class__.__name__,
-        )
-        return remove_columns_collator
-
-    def _is_iterable_dataset(self, dataset):
-        return isinstance(dataset, paddle.io.IterableDataset)
-
-    def print_config(self, args=None, key=""):
-        """
-        print config values
-        """
-        logger.info("=" * 60)
-        if args is None:
-            args = self.args
-            key = "Training"
-        import paddlenlp
-
-        logger.info("{:^40}".format("{} Configuration Arguments".format(key)))
-        logger.info("{:30}: {}".format("paddle commit id", paddle.version.commit))
-        logger.info("{:30}: {}".format("paddlenlp commit id", paddlenlp.version.commit))
-
-        for a in dir(args):
-            if a[:2] != "__":  # don't print double underscore methods
-                v = getattr(args, a)
-                if not isinstance(v, types.MethodType):
-                    logger.info("{:30}: {}".format(a, v))
-
-        logger.info("")
